Move tests into similar test classes
