--- embrace-android-sdk/src/integrationTest/kotlin/io/embrace/android/embracesdk/testcases/session/BackgroundActivityTest.kt
@@ -1,50 +1,0 @@
-package io.embrace.android.embracesdk.testcases.session
-
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import io.embrace.android.embracesdk.assertions.getSessionPartId
-import io.embrace.android.embracesdk.assertions.hasSpanSnapshotsOfType
-import io.embrace.android.embracesdk.internal.arch.schema.EmbType
-import io.embrace.android.embracesdk.internal.config.remote.BackgroundActivityRemoteConfig
-import io.embrace.android.embracesdk.internal.config.remote.RemoteConfig
-import io.embrace.android.embracesdk.internal.arch.state.AppState
-import io.embrace.android.embracesdk.testframework.SdkIntegrationTestRule
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertFalse
-import org.junit.Assert.assertNotEquals
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Asserts that a stateful background activity can be recorded.
- */
-@RunWith(AndroidJUnit4::class)
-internal class BackgroundActivityTest {
-
-    @Rule
-    @JvmField
-    val testRule: SdkIntegrationTestRule = SdkIntegrationTestRule()
-
-    @Test
-    fun `bg activity messages are recorded`() {
-        testRule.runTest(
-            persistedRemoteConfig = RemoteConfig(backgroundActivityConfig = BackgroundActivityRemoteConfig(100f)),
-            testCaseAction = {
-                recordSession()
-                clock.tick(30000)
-                recordSession()
-            },
-            assertAction = {
-                // filter out dupes from overwritten saves
-                val bgActivities = getSessionEnvelopes(2, AppState.BACKGROUND).distinctBy { it.getSessionPartId() }
-                assertEquals(2, bgActivities.size)
-
-                val first = bgActivities[0]
-                assertFalse(first.hasSpanSnapshotsOfType(EmbType.Ux.Session))
-
-                // ID should be different for each
-                assertNotEquals(first.getSessionPartId(), bgActivities[1].getSessionPartId())
-            }
-        )
-    }
-}
