+++ embrace-android-sdk/src/integrationTest/kotlin/io/embrace/android/embracesdk/testcases/session/ManualUserSessionTest.kt
@@ -3,11 +3,18 @@ package io.embrace.android.embracesdk.testcases.session
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.embrace.android.embracesdk.assertions.findSessionSpan
 import io.embrace.android.embracesdk.assertions.getUserSessionId
+import io.embrace.android.embracesdk.assertions.getUserSessionTerminationReason
+import io.embrace.android.embracesdk.assertions.isFinalSessionPart
 import io.embrace.android.embracesdk.internal.config.remote.BackgroundActivityRemoteConfig
 import io.embrace.android.embracesdk.internal.config.remote.RemoteConfig
 import io.embrace.android.embracesdk.internal.config.remote.SessionRemoteConfig
+import io.embrace.android.embracesdk.semconv.EmbSessionAttributes.EmbUserSessionTerminationReasonValues.MANUAL
 import io.embrace.android.embracesdk.testframework.SdkIntegrationTestRule
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotEquals
+import org.junit.Assert.assertNull
+import org.junit.Assert.assertTrue
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -89,6 +96,54 @@ internal class ManualUserSessionTest {
                 val fgSessions = getSessionEnvelopes(2)
                 assertNotEquals(fgSessions[0].getUserSessionId(), fgSessions[1].getUserSessionId())
             },
+        )
+    }
+
+    @Test
+    fun `new user session can be manually created`() {
+        testRule.runTest(
+            testCaseAction = {
+                recordSession {
+                    clock.tick(10_000)
+                    embrace.endUserSession()
+                }
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(2)
+
+                assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+                assertTrue(sessions[0].isFinalSessionPart())
+                assertEquals(MANUAL, sessions[0].getUserSessionTerminationReason())
+                assertFalse(sessions[1].isFinalSessionPart())
+                assertNull(sessions[1].getUserSessionTerminationReason())
+            }
+        )
+    }
+
+    @Test
+    fun `endUserSession past the manual cooldown window ends a second user session`() {
+        testRule.runTest(
+            testCaseAction = {
+                recordSession {
+                    // user session old enough for the first manual end to take effect
+                    clock.tick(10_000)
+                    embrace.endUserSession()
+                    // past the 5s cooldown, so the second manual end also takes effect
+                    clock.tick(6_000)
+                    embrace.endUserSession()
+                }
+            },
+            assertAction = {
+                val sessions = getSessionEnvelopes(3)
+                assertNotEquals(sessions[0].getUserSessionId(), sessions[1].getUserSessionId())
+                assertNotEquals(sessions[1].getUserSessionId(), sessions[2].getUserSessionId())
+                assertTrue(sessions[0].isFinalSessionPart())
+                assertEquals(MANUAL, sessions[0].getUserSessionTerminationReason())
+                assertTrue(sessions[1].isFinalSessionPart())
+                assertEquals(MANUAL, sessions[1].getUserSessionTerminationReason())
+                assertFalse(sessions[2].isFinalSessionPart())
+                assertNull(sessions[2].getUserSessionTerminationReason())
+            }
         )
     }
 }
